+++ tests/e2e/core-tests/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Unreleased
 
+# 0.1.1
+
 ## Added
+
+- Registered Shopper Checkout tests
 - Merchant Order Status Filter tests
 - Merchant Order Refund tests
 - Merchant Apply Coupon tests
@@ -8,6 +12,9 @@
 - Merchant Product Edit tests
 - Merchant Product Search tests
 - Shopper Checkout Apply Coupon
+
+
+- Shopper Cart Apply Coupon
 
 ## Fixed